moved _approximte_mode in utils
